+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 		* Supports adding standalone mythic spellbooks with a CL defined by a context value.
 	* IncreaseActivatableAbilitySpeed
 		* Supports altering command speed of activatable abiltiies.
+	* ContextConditionCasterHasResource
+		* Supports checking a resource on the caster.
 * New Events
 	* IActivatableAbilityGetCommandTypeHandler
 		* Allows the altering of activatable ability command types at run time arbitrarily.